alpacahack/daily/hit-and-miss/pyproject.toml
[project]
name = "hit-and-miss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pwntools>=4.15.0",
]
